assignment_8/requirements.txt
boofuzz
setuptools
python-dotenv
pwntools
ipykernel
jupyter